+++ lua/ui/lobby/changelogData.lua
@@ -12,6 +12,24 @@ last_version = 3747
 ---@type PatchNotes[]
 gamePatches = {
     {
+        version = 3748,
+        name = 'Hotfix',
+        description = {
+            "",
+            "## Bug fixes",
+            "",
+            "- (#4414) Revert changes to prefetching of assets",
+            "    We thought we understood how prefetching worked, but we were wrong. As a result some",
+            "    people ran out of memory. That results in an application crash.",
+            "",
+            "    We reverted the changes.",
+            "",
+            "## Contributors",
+            "",
+            "- Jip (#4414)",
+        }
+    },
+    {
         version = 3747,
         name = 'Hotfix',
         description = {